Float64 until next jll

+++ simulations/atmosphere_simulation_compile.jl
@@ -6,7 +6,7 @@ using Oceananigans
 using Oceananigans.Architectures: ReactantState
 
 PROFILE[] = true
-Oceananigans.defaults.FloatType = Float32
+Oceananigans.defaults.FloatType = Float64
 
 preamble()
